es filter methods to sfeos helpers
